requirements.txt
oracledb==1.4.1
python-dotenv==1.0.0
tabulate==0.9.0
asyncio==3.4.3 